// carbon-credit-webapp/src/atoms/CCTable/CCTable.interface.ts
export interface CCTableProps {
  headings: any
  rows?: any[]
  maxWidth?: any
  pagination?: boolean
  tableSx?: any
  sx?: any
  loading?: boolean
  data?: boolean
  rowsPerPageProp?: number
  hideScrollbar?: boolean
  lastTwoColsSticky?: boolean
  stickyLastCol?: boolean
  stickySecondLastCol?: boolean
}